Fix OIIO build against OpenJPH for 2025
src/jpeg2000.imageio/CMakeLists.txt assumes the includes for
openjph are in the same location as openjpeg which is not the
case when building in Conan.

This doesn't show up in 2026 since openexr 3.4.x declares a dependency
on openjph and since OIIO has a dependency on openexr, it gets the
include path that way.

Signed-off-by: Jean-Francois Panisset <panisset@gmail.com>